testing

+++ src/Connector J 8.0/finalProject.java
@@ -814,7 +814,7 @@ class finalProject {
                 rs = ps.getResultSet();
                 //rs.beforeFirst();
                 while (rs.next()) {
-                    System.out.println(rs.getInt(1) + " , " + rs.getString(2) + " , " + rs.getString(3) + " , " + rs.getString(4) + ", " + rs.getInt(5));
+                    System.out.println(rs.getInt(1) + " , " + rs.getString(2) + " , " + rs.getString(3) + " , " + rs.getString(4) + ", " + rs.getDouble(5));
                 }
 
             }
